Cambios finales

+++ app/Http/Livewire/UsersLivewire.php
@@ -11,53 +11,53 @@ class UsersLivewire extends Component
 
 {
 
-    public $name ='' ;
-    public $email ='' ;
-    public $password ='' ;
-    public $password_confirmation='' ;
+    public $name = '';
+    public $email = '';
+    public $password = '';
+    public $password_confirmation = '';
     public $actualizando;
 
-    public function createUser(){
-       
-            User::create([
-                'name' => $this->name,
-                'email' => $this->email,
-                'password' => Hash::make($this->password),
-            ]);
-    
-            // Limpiar los campos después de guardar
-            $this->reset();
-            session()->flash('message', 'Usuario creado exitosamente.');
-       
-        }
+    public function createUser()
+    {
 
-    
+        User::create([
+            'name' => $this->name,
+            'email' => $this->email,
+            'password' => Hash::make($this->password),
+        ]);
 
-    public function delete($id){
+        // Limpiar los campos después de guardar
+        $this->reset();
+        session()->flash('message', 'Usuario creado exitosamente.');
+    }
+
+
+
+    public function delete($id)
+    {
 
         User::destroy($id);
         session()->flash('error', 'Usuario eliminado correctamente .');
-
     }
 
-    
+
     public function edit($id)
     {
-        $user= User::find($id);
+        $user = User::find($id);
         $this->name = $user->name;
         $this->email = $user->email;
         $this->password = $user->password;
- 
+
 
         $this->actualizando = true; // Cambia a modo actualización
-        session()->flash('message', ' El Id del usuario que va actualizar es el  .'.$id);
+        session()->flash('message', ' El Id del usuario que va actualizar es el  .' . $id);
     }
 
     public function update($id)
     {
         $user = User::findOrFail($id);
         // Si se ha subido una nueva imagen, se guarda y se actualiza la URL de la imagen
-      
+
 
         // Actualizar los demás campos del producto
         $user->name = $this->name;
@@ -66,10 +66,9 @@ class UsersLivewire extends Component
         $user->save();
         $this->reset();
         session()->flash('message', 'Usuario actualizado exitoxamente .');
-        
     }
 
-  //codigo del modal para eliminar  un usuario
+    //codigo del modal para eliminar  un usuario
     public $modalAbierto = false;
     public $contrasena = '';
     public $mensajeError = '';
@@ -78,7 +77,7 @@ class UsersLivewire extends Component
     public function abrirModal($id)
     {
         $this->modalAbierto = true;
-        $this->UsuarioEliminar=$id;
+        $this->UsuarioEliminar = $id;
     }
 
     public function cerrarModal()
@@ -88,22 +87,26 @@ class UsersLivewire extends Component
 
     public function eliminarUsuario()
     {
-           // Obtener el usuario autenticado
-           $user = Auth::user();
+        // Obtener el usuario autenticado
+        $user = Auth::user();
+        //verifica que no se elimine el usuario logeado
+        if ($this->UsuarioEliminar != $user->id) {
+            // Verificar si la contraseña es correcta
+            if (password_verify($this->contrasena, $user->password)) {
+                // Eliminar el usuario
+                User::findOrFail($this->UsuarioEliminar)->delete();
 
-           // Verificar si la contraseña es correcta
-           if (password_verify($this->contrasena, $user->password)) {
-               // Eliminar el usuario
-               User::findOrFail($this->UsuarioEliminar)->delete();
-   
-               // Cerrar el modal después de eliminar el usuario
-               $this->cerrarModal();
-               session()->flash('message', 'Usuario eliminado exitoxamente .');
-           } else {
-               // Mostrar un mensaje de error si la contraseña es incorrecta
-               
-               session()->flash('message', 'Contraseña incorrecta .');
-           }
+                // Cerrar el modal después de eliminar el usuario
+                $this->cerrarModal();
+                session()->flash('message', 'Usuario eliminado exitoxamente .');
+            } else {
+                // Mostrar un mensaje de error si la contraseña es incorrecta
+
+                session()->flash('message', 'Contraseña incorrecta .');
+            }
+        }else {
+            session()->flash('message', 'No te puedes elimiar a ti mismo .');
+        }
     }
